python/requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
psycopg2-binary==2.8.6
flask-login==0.5.0